feat: Temporary checkpoint for emulator

+++ cpue/include/loader.h
@@ -21,7 +21,7 @@ struct Region {
     u64 flags = REGION_WRITABLE;
     u8* data = nullptr;
 
-    [[nodiscard]] u64 get_page_count() const { return std::ceil(size / PAGE_SIZE); }
+    [[nodiscard]] u64 get_page_count() const { return std::ceil(static_cast<double>(size) / PAGE_SIZE); }
 };
 
 class Loader {
